move details component into respective folder
