+++ src/editor/details/info-details/info-details.component.spec.ts
