+++ src/editor/details/service/details.service.ts
@@ -7,7 +7,7 @@ import {
 } from "../information-link-details/information-link-details.component";
 import {ConversationPartner} from "../../../shared/models/keml/conversation-partner";
 import {ConversationPartnerDetailsComponent} from "../cp-details/cp-details.component";
-import {InfoDetailsComponent} from "../../info-details/info-details.component";
+import {InfoDetailsComponent} from "../info-details/info-details.component";
 
 @Injectable({
   providedIn: 'root'
