+++ src/editor/editor.module.ts
@@ -12,7 +12,7 @@ import {InfoComponent} from "./info/info.component";
 import {DatabaseSvgComponent} from "./database-svg/database-svg.component";
 import {MsgDetailsComponent} from "./details/msg-details/msg-details.component";
 import {ConversationPartnerDetailsComponent} from "./details/cp-details/cp-details.component";
-import {InfoDetailsComponent} from "./info-details/info-details.component";
+import {InfoDetailsComponent} from "./details/info-details/info-details.component";
 import {IsInstrSvgComponent} from "./is-instr-svg/is-instr-svg.component";
 import {PreknowledgeComponent} from "./preknowledge/preknowledge.component";
 import {TextAreaSvgComponent} from "./text-area-svg/text-area-svg.component";
